feats
improved quality and logic. Also finished most of the assignments

+++ BetterRecipe/BetterRecipe/List/RecipeList.swift
@@ -1,10 +1,3 @@
-//
-//  RecipeList.swift
-//  BetterRecipe
-//
-//  Created by Catalin Posedaru on 15/12/25.
-//
-
 import SwiftUI
 import SwiftData
 
@@ -14,10 +7,8 @@ struct RecipeList: View {
     @State private var isSorted = false
     @State private var newRecipe: Recipe?
     @State private var sortType: String = ""
-    @State private var sortedRecipes: [Recipe]?
 
     init(searchText: String = "") {
-        sortedRecipes = nil
         let predicate = #Predicate<Recipe> { recipe in
             searchText.isEmpty || recipe.title
                 .localizedStandardContains(searchText)
@@ -34,10 +25,18 @@ struct RecipeList: View {
     }
     
     private func deleteRecipe(indexes: IndexSet) {
+        var toDelete: Recipe?
         for index in indexes {
-            context.delete(recipes[index])
+            toDelete = sortedRecipes[index]
         }
-        
+        if let toDelete = toDelete {
+            for index in 0..<recipes.count {
+                if recipes[index].title == toDelete.title {
+                    context.delete(recipes[index])
+                }
+            }
+        }
+
         if context.hasChanges {
             saveContext()
         }
@@ -52,22 +51,22 @@ struct RecipeList: View {
         )
     }
 
-    private func recipeSorter() {
+    var sortedRecipes: [Recipe] {
         if sortType == "Favorite" {
             withAnimation(.easeInOut) {
-                sortedRecipes = recipes.filter {
+                return recipes.filter {
                     $0.isFavorite
                 }
             }
         } else if (sortType == "Alphabetical"){
             withAnimation(.easeInOut) {
-                sortedRecipes = recipes.sorted { first, second in
+                return recipes.sorted { first, second in
                     first.title < second.title
                 }
             }
         } else {
             withAnimation(.easeInOut) {
-                sortedRecipes = recipes
+                return recipes
             }
         }
     }
@@ -78,57 +77,13 @@ struct RecipeList: View {
             VStack {
                 TitleView()
 
+                listView
+
                 Rectangle()
                     .fill(.darkerGreen)
                     .frame(height: 2)
 
-                List {
-                    ForEach(
-                        recipes
-                    ) { recipe in
-                        NavigationLink(
-                            destination: RecipeDetail(
-                                recipe: recipe,
-                                editRecipe: false,
-                                newRecipe: false
-                            )
-                            .background(.softGreen.opacity(0.7))
-                        ) {
-                            Text(recipe.title)
-                                .font(.system(size: 22))
-                            Spacer()
-                            Button("Favorite", systemImage: !recipe.isFavorite ? "star" : "star.fill") {
-                                recipe.isFavorite.toggle()
-                                saveContext()
                             }
-                            .font(.title)
-                            .labelStyle(.iconOnly)
-                            .tint(.orangeAccent)
-                            .buttonStyle(.borderless)
-                        }
-                        .navigationLinkIndicatorVisibility(.hidden)
-                    }
-                    .onDelete(perform: deleteRecipe(indexes:))
-                    .listRowSeparatorTint(.black)
-                    .listRowBackground(Color.darkerGreen.opacity(0.4))
-                }
-                .listStyle(.insetGrouped)
-                .scrollContentBackground(.hidden)
-                .toolbar {
-                    ToolbarItem (placement: .topBarLeading) {
-                        Button("Add", systemImage: "plus") {
-                            addRecipe()
-                        }
-                    }
-                    ToolbarItem (placement: .topBarTrailing) {
-                        Button("Sort by...") {
-                            isSorted = true
-                        }
-                        .font(.system(size: 20))
-                        .fontWeight(.medium)
-                    }
-                }
-            }
             .background(.softGreen.opacity(0.7))
             .confirmationDialog(
                 "Select a sorting method",
@@ -136,23 +91,69 @@ struct RecipeList: View {
             ) {
                 Button("Alphabetical") {
                     sortType = "Alphabetical"
-                    recipeSorter()
                 }
-                
+
                 Button("Favorites") {
                     sortType = "Favorite"
-                    recipeSorter()
                 }
             }
             .sheet(item: $newRecipe) { recipe in
                 RecipeDetail(
                     recipe: recipe,
-                    editRecipe: true,
-                    newRecipe: true)
+                    state: .creating)
                 .background(.softGreen.opacity(0.7))
             }
 
         }
+    }
+
+    var listView: some View {
+        List {
+            ForEach(
+                sortedRecipes
+            ) { recipe in
+                NavigationLink(
+                    destination: RecipeDetail(
+                        recipe: recipe,
+                        state: .reading
+                    )
+                    .background(.softGreen.opacity(0.7))
+                ) {
+                    Text(recipe.title)
+                        .font(.system(size: 22))
+                    Spacer()
+                    Button("Favorite", systemImage: !recipe.isFavorite ? "star" : "star.fill") {
+                        recipe.isFavorite.toggle()
+                        saveContext()
+                    }
+                    .font(.title)
+                    .labelStyle(.iconOnly)
+                    .tint(.orangeAccent)
+                    .buttonStyle(.borderless)
+                }
+                .navigationLinkIndicatorVisibility(.hidden)
+            }
+            .onDelete(perform: deleteRecipe(indexes:))
+            .listRowSeparatorTint(.black)
+            .listRowBackground(Color.darkerGreen.opacity(0.4))
+        }
+        .listStyle(.insetGrouped)
+        .scrollContentBackground(.hidden)
+        .toolbar {
+            ToolbarItem (placement: .topBarLeading) {
+                Button("Add", systemImage: "plus") {
+                    addRecipe()
+                }
+            }
+            ToolbarItem (placement: .topBarTrailing) {
+                Button("Sort by...") {
+                    isSorted = true
+                }
+                .font(.system(size: 20))
+                .fontWeight(.medium)
+            }
+        }
+
     }
 }
